Moved insert comment in own class in util section
